* Implement VOIKKO_OPT_COMPOUND_INTERSECT_LEVEL
git-svn-id: https://voikko.svn.sourceforge.net/svnroot/voikko/trunk@84 ee56993d-3414-0410-8ed4-8136e77a0dd8

+++ libvoikko/src/voikko_setup.c
@@ -102,7 +102,7 @@ int voikko_terminate(int handle) {
 	else return 0;
 }
 
-int voikko_find_malaga_project(char * buffer, int buflen, const char * langcode) {
+int voikko_find_malaga_project(char * buffer, size_t buflen, const char * langcode) {
 	struct passwd pwd;
 	struct stat sbuf;
 	struct passwd * pwd_result;
